[meditor-vis] Upload first version of the tool to create visualizations from quality models
The tool create a dashboard for each attribute in the quality model showing the metrics
for this attribute. Sample execution:

PYTHONPATH=. meditor/meditor-vis.py -e elastic_url -t "AttributeTemplate" -m "Developer Quality Model" -i ossmeter

From the "Developer Quality Model" in meditor and using the metrics data in the "ossmeter" index
in the Elasticsearch in elastic_url, a new dashboard is created in the Kibana in elastic_url
using the template "AttributeTemplate" which must be already uploaded to Kibana
using kidash.

The dashboard is created from an already created attribute template
dashboard: attribute-template.json.