[prosoul_assess] Store empty values to ES
This code allows to store the empty values obtained from
the assessment of the quality models in ES. After calculating
the assessment dict, all projects available in ES are retrieved
and used to calculate the diff of the assessment dict. It mimics
the structure of the assessment dict and it includes only the
empty metrics. This operation is applied for quarters and
over all the time span of the assessment.

The empty data is stored in ES, in two different indexes. One
contains the empty data for quarters, while the other one
the empty data for all the time span. Indexes with data and
with empty data share the same alias, thus allowing to
perform searches on both type of data in a transparent way.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>